Update regex_example1.cs
The code would currently return that $3, or any other integer currency <10, is not a currency. This is because the regex is always looking for a match of 2 decimal digits in '\d{2}?' regardless of whether a decimal point was found. The change checks for zero or one instance(s) of the decimal match '(\.?\d{2}?)'.